requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.2
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.4
sentence-transformers==2.2.2
azure-storage-blob==12.19.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6 
huggingface_hub==0.24.0